fix bugs

+++ tg_bot/models.py
@@ -6,7 +6,7 @@ from sqlalchemy.schema import UniqueConstraint
 class UserModel(db.Model):
     __tablename__ = 'user'
     id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, unique=True, nullable=False)
+    user_id = db.Column(db.BigInteger, unique=True, nullable=False)
     first_name = db.Column(db.String(20))
     last_name = db.Column(db.String(40))
     username = db.Column(db.String(30))
